feat: stageManager.cs
StageController를 StageManager로 이름 변경 하였음
